requirements.txt
gpiozero
pychalk
python-dotenv